added razorpay integration flow for payment

+++ src/main/java/com/uber/bookingApp/model/RideRequest.java
@@ -39,7 +39,7 @@ public class RideRequest {
     @Enumerated(EnumType.STRING)
     private RideRequestStatus requestStatus;
 
-    private Double fare;
+    private Long fare;
 
 
 }
